Maze AI v0.5.0 MazeServer.MazeAi

this is an AI for maze game that developed with elixir it solve maze with this algorithms:

Breadth-first search (BFS)

Iterative deepening depth-first search (IDS)

A search (A Star) all of this algorithms will use graph search algorithm with * special frontier rules or heuristical function.

Functions

create_point(point, board, g, h, end_point, parent)

it takes a location in board (x, y) and make a point node.

expand(point, board, frontier, explored_set, frontier_push, g, h, end_point, expanding)

expand function will create child nodes.

expander(frontier, point, explored_set, frontier_push)

duty of expander is check frontier queue and explored set for avoiding of redundancy and revisiting a node and visiting a wall node. it will return a new valid frontier.

Examples

iex> MazeServer.MazeAi.expander([%{x: 1, y: 2}], %{x: 1, y: 2, state: "0"}, [%{x: 4, y: 3}], fn f, p -> List.insert_at(f, -1, p) end)
[%{x: 1, y: 2}]

iex> MazeServer.MazeAi.expander([%{x: 2, y: 4}], %{x: 1, y: 2, state: "0"}, [%{x: 1, y: 2}], fn f, p -> List.insert_at(f, -1, p) end)
[%{x: 2, y: 4}]

iex> MazeServer.MazeAi.expander([%{x: 1, y: 2}], %{x: 4, y: 2, state: "1"}, [%{x: 4, y: 3}], fn f, p -> List.insert_at(f, -1, p) end)
[%{x: 1, y: 2}]

iex> MazeServer.MazeAi.expander([%{x: 1, y: 2}], %{x: 4, y: 2, state: "0"}, [%{x: 4, y: 3}], fn f, p -> List.insert_at(f, -1, p) end)
[%{x: 1, y: 2}, %{state: "0", x: 4, y: 2}]

find_target(board)

find_target takes a board, search for target point in it then return end_point.

Examples

iex> MazeServer.MazeAi.find_target(["111", "121", "111"])
{1, 1}

graph_search(frontier, explored_set, board, goal, wall, limit, frontier_pop, frontier_push)

graph_search function is a basical function for search on graphs. it use frontier_push and frontier_pop for define frontier rules.

graph_search(frontier, explored_set, end_point, board, goal, wall, limit, g, h, frontier_pop, frontier_push, expander \\ &expander/4)

init_board()

this function will return main test board.