Maze AI v0.1.0 MazeServer.MazeAi.AStar

this module define A* rules. in this algorithm, frontier is same as BFS with a tiny difference! frontier will sort by ascending path_cost order after every push.

Functions

expander(frontier, point, explored_set, frontier_push)

this expander is same as MazeAi.expander but it will check path_cost for expand a node!

frontier_pop(frontier)

frontier pop rule for A*.

frontier_push(frontier, point)

frontier push rule for A*. it will sort frontier by ascending path_cost order.

g(path_cost, _)

it will calculates path cost of a node with path cost of its parent node plus one!

h(arg1, arg2)

heuristic calculation with diagonal distance calculator.

Examples

iex> MazeServer.MazeAi.AStar.h({1, 2}, {4, 5})
6

and this is true: 2 * max((4-1), (5-2)) = 6

search(board \\ MazeAi.init_board(), point \\ %{x: 1, y: 14})

A* search. it will search same as BFS.