Maze AI v0.1.0 MazeServer.MazeAi.BFS

this module define BFS rules. in this algorithm, frontier is same as queue! (FIFO queue)

Functions

frontier_pop(frontier)

frontier poh rule for BFS this function will remove first node of list. (First Out)

frontier_push(frontier, point)

frontier push rule for BFS this function will add any new node in the last of list. (First In)

search(board \\ MazeAi.init_board(), point \\ %{x: 1, y: 14})

BFS search function.